requirements.txt
flask
opencv-python
face-recognition
numpy
mysql-connector-python
